+++ KSP2Runtime/KSPOrbit/KSPOrbitModule.Waypoint.cs
@@ -1,0 +1,5 @@
+﻿namespace KontrolSystem.KSP.Runtime.KSPMission;
+
+public class Waypoint {
+    
+}